% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_cancer}
\alias{breast_cancer}
\title{The breast cancer wisconsin dataset.}
\format{
A data frame with 569 rows and 31 variables (30 covariates):
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Diagnostic)}
}
\usage{
breast_cancer
}
\description{
The breast cancer wisconsin dataset for binary classification (benign or malignant)
}
\keyword{datasets}
