% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digits}
\alias{digits}
\title{The digits dataset.}
\format{
A data frame with 1797 rows and 65 variables (64 covariates):
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Optical+Recognition+of+Handwritten+Digits}
}
\usage{
digits
}
\description{
The digits dataset for multi-class classification (handwritten digits recognition).
}
\keyword{datasets}
