% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Do K-means clustering}
\usage{
get_clusters(x, centers = 2L, seed = 123L, clustering_obj = NULL)
}
\arguments{
\item{x}{a numeric matrix(like object) of predictors}

\item{centers}{number of clusters}

\item{seed}{random seed for reproducibility}

\item{clustering_obj}{a list of kmeans results. Default is NULL, at training time.
Must be provided at prediction time.}
}
\value{
a list of \code{\link{kmeans}} results, with additional
attributes: xm, xsd, encoded_x
}
\description{
Do K-means clustering
}
\examples{

n <- 7 ; p <- 3

X <- matrix(rnorm(n * p), n, p) # no intercept!

print(get_clusters(X))


}
