% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_predict_bcn.R
\name{predict.bcn}
\alias{predict.bcn}
\title{Predict method for Boosted Configuration Networks (BCN)}
\usage{
\method{predict}{bcn}(object, newx, type = c("response", "probs"))
}
\arguments{
\item{object}{a object of class 'bcn'}

\item{newx}{new data, with no intersection with training data}

\item{type}{a string, "response" is the class, "probs" are the classifier's probabilities}
}
\description{
Predict method for Boosted Configuration Networks (BCN)
}
\examples{

set.seed(1234)
train_idx <- sample(nrow(iris), 0.8 * nrow(iris))
X_train <- as.matrix(iris[train_idx, -ncol(iris)])
X_test <- as.matrix(iris[-train_idx, -ncol(iris)])
y_train <- iris$Species[train_idx]
y_test <- iris$Species[-train_idx]

fit_obj <- bcn::bcn(x = X_train, y = y_train, B = 10, nu = 0.335855,
lam = 10**0.7837525, r = 1 - 10**(-5.470031), tol = 10**-7,
activation = "tanh", type_optim = "nlminb")

print(predict(fit_obj, newx = X_test) == y_test)
print(mean(predict(fit_obj, newx = X_test) == y_test))

print(predict(fit_obj, newx = X_test, type="probs"))

}
