% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_optimization_functions.R
\name{random_search}
\alias{random_search}
\title{Random Search}
\usage{
random_search(
  objective,
  lower,
  upper,
  seed = 123,
  control = list(iter.max = 100)
)
}
\arguments{
\item{objective}{objective function to be minimized}

\item{lower}{lower bound for search}

\item{upper}{upper bound for search}

\item{seed}{an integer, for reproducing the result}

\item{control}{a list of control parameters. For now \code{control = list(iter.max=100)},
where \code{iter.max} is the maximum number of iterations allowed}
}
\value{
A list with components

\code{par}	the best set of parameters found

\code{objective}	the value of objective corresponding to par

\code{iterations} number of iterations performed
}
\description{
Random Search derivative-free optimization
}
\examples{

fr <- function(x) {   ## Rosenbrock Banana function
x1 <- x[1]
x2 <- x[2]
100 * (x2 - x1 * x1)^2 + (1 - x1)^2
}

random_search(fr, lower = c(-2, -2), upper = c(2, 2), control = list(iter.max=1000))

}
