% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{The wine dataset.}
\format{
A data frame with 178 rows and 14 variables (13 covariates):
}
\source{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/wine/wine.data}
}
\usage{
wine
}
\description{
The wine dataset for multi-class classification.
}
\keyword{datasets}
